Move test files into a folder
