+++ __tests__/Auth0Client/constructor.ts
@@ -1,15 +1,15 @@
 import 'fast-text-encoding';
 import unfetch from 'unfetch';
-import { verify } from '../src/jwt';
+import { verify } from '../../src/jwt';
 import { MessageChannel } from 'worker_threads';
-import * as utils from '../src/utils';
-import * as scope from '../src/scope';
+import * as utils from '../../src/utils';
+import * as scope from '../../src/scope';
 
 // @ts-ignore
 
-import { assertUrlEquals, setupFn } from './Auth0Client.helpers';
+import { setupFn } from './helpers';
 
-import { TEST_CLIENT_ID, TEST_CODE_CHALLENGE, TEST_DOMAIN } from './constants';
+import { TEST_CLIENT_ID, TEST_CODE_CHALLENGE, TEST_DOMAIN } from '../constants';
 
 jest.mock('unfetch');
 jest.mock('es-cookie');
@@ -25,6 +25,15 @@ jest
   .mockReturnValue(TEST_CODE_CHALLENGE);
 
 jest.spyOn(utils, 'runPopup');
+
+const assertUrlEquals = (actualUrl, host, path, queryParams) => {
+  const url = new URL(actualUrl);
+  expect(url.host).toEqual(host);
+  expect(url.pathname).toEqual(path);
+  for (let [key, value] of Object.entries(queryParams)) {
+    expect(url.searchParams.get(key)).toEqual(value);
+  }
+};
 
 const setup = setupFn(mockVerify);
 
@@ -66,6 +75,61 @@ describe('Auth0Client', () => {
     mockFetch.mockReset();
     jest.clearAllMocks();
     window.location = oldWindowLocation;
+  });
+
+  describe('constructor', () => {
+    it('automatically adds the offline_access scope during construction', () => {
+      const auth0 = setup({
+        useRefreshTokens: true,
+        scope: 'test-scope'
+      });
+
+      expect((<any>auth0).scope).toBe('test-scope offline_access');
+    });
+
+    it('ensures the openid scope is defined when customizing default scopes', () => {
+      const auth0 = setup({
+        advancedOptions: {
+          defaultScope: 'test-scope'
+        }
+      });
+
+      expect((<any>auth0).defaultScope).toBe('openid test-scope');
+    });
+
+    it('allows an empty custom default scope', () => {
+      const auth0 = setup({
+        advancedOptions: {
+          defaultScope: null
+        }
+      });
+
+      expect((<any>auth0).defaultScope).toBe('openid');
+    });
+
+    it('should create issuer from domain', () => {
+      const auth0 = setup({
+        domain: 'test.dev'
+      });
+
+      expect((<any>auth0).tokenIssuer).toEqual('https://test.dev/');
+    });
+
+    it('should allow issuer as a domain', () => {
+      const auth0 = setup({
+        issuer: 'foo.bar.com'
+      });
+
+      expect((<any>auth0).tokenIssuer).toEqual('https://foo.bar.com/');
+    });
+
+    it('should allow issuer as a fully qualified url', () => {
+      const auth0 = setup({
+        issuer: 'https://some.issuer.com/'
+      });
+
+      expect((<any>auth0).tokenIssuer).toEqual('https://some.issuer.com/');
+    });
   });
 
   describe('buildLogoutUrl', () => {
