+++ __tests__/Auth0Client/handleRedirectCallback.test.ts
@@ -1,24 +1,20 @@
 import 'fast-text-encoding';
 import unfetch from 'unfetch';
-import { verify } from '../src/jwt';
+import { verify } from '../../src/jwt';
 import { MessageChannel } from 'worker_threads';
-import * as utils from '../src/utils';
-import * as scope from '../src/scope';
+import * as utils from '../../src/utils';
+import * as scope from '../../src/scope';
 
 // @ts-ignore
 
-import {
-  fetchResponse,
-  loginWithRedirectFn,
-  setupFn
-} from './Auth0Client.helpers';
+import { fetchResponse, loginWithRedirectFn, setupFn } from './helpers';
 
 import {
   TEST_ACCESS_TOKEN,
   TEST_CODE_CHALLENGE,
   TEST_ID_TOKEN,
   TEST_REFRESH_TOKEN
-} from './constants';
+} from '../constants';
 
 jest.mock('unfetch');
 jest.mock('es-cookie');
